requirements.txt
python-telegram-bot==21.4
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
